requirements.txt
altgraph==0.17.5
annotated-types==0.7.0
et_xmlfile==2.0.0
greenlet==3.3.2
numpy==2.4.3
openpyxl==3.1.5
packaging==26.0
pandas==3.0.1
pefile==2024.8.26
pillow==12.1.1
pydantic==2.12.5
pydantic_core==2.41.5
pyinstaller==6.19.0
pyinstaller-hooks-contrib==2026.3
python-dateutil==2.9.0.post0
pywin32-ctypes==0.2.3
setuptools==82.0.1
six==1.17.0
SQLAlchemy==2.0.48
sqlmodel==0.0.37
ttkbootstrap==1.20.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
